Add type hints to fab.util
This adds type hints to the various functions in fab.util.  The only
code change involves initialising Timer.taken to 0.0 because this
should be set by the time it is used in TimeLogger but the previous
type hints allowed for it to be None, upsetting the checker.

The return type of __entry__ methods needs to be the class name
as a string unless annotations is imported from __future__
because it makes a forward reference to the class being defined.
The types of __exit__ method arguments are based on information
available online.

The config argument to input_to_output_fpath is currently ignored
because it should be an instance of BuildConfig, but importing
the class from build_config appears to introduce a circular
library dependency.