Add type hinting to easy cases
This groups a handful for type hinting changes to three modules which
require minimal changes to function definitions.

The only minor issue is the use of an ignore in metrics to skip checks
on key type hintign in an in-line sort.  It may be possible to revisit
this when the rest of fab has been typed.