Adopt subprocess fixture for testing. (#391)
Rather than variations of the same hand-mocking of shelling out to a
tool this change takes advantage of an existing fixture.

It also pulls in a fixture to create a fake filing system in memory
which should be faster and less prone to faults based on what the
developer has in their file tree.